Revert "Revert "lint: remove unnecessary type conversions""
This reverts commit fe4d7771d9de759d15973243d40b80f10df0054d.

+++ client/go/internal/vespa/document/http.go
@@ -352,6 +352,6 @@ func (c *Client) resultWithResponse(resp *http.Response, sentBytes int, result R
 	}
 	result.Latency = elapsed
 	result.BytesSent = int64(sentBytes)
-	result.BytesRecv = int64(written)
+	result.BytesRecv = written
 	return result
 }
